Add files via upload
 I had successfully completed my Level 1 #task3 project to create a Temperature Converter using HTML, CSS, JAVASCRIPT during my internship at Oasis Infobyte is a valuable accomplishment. Building a temperature converter is a common exercise in web development, and it's a fantastic way to apply your skills in HTML, CSS, and JavaScript. By completing this task, I had likely gained valuable experience in working with user interfaces, handling input and output, and implementing logic in JavaScript.
A temperature converter typically allows users to convert temperatures between different units, such as Celsius and Fahrenheit. It involves taking user input, performing the necessary calculations, and displaying the converted result.
It's worth noting that HTML provides the structure and layout of webpage, CSS adds styling and visual enhancements, and JavaScript handles the logic and interactivity of temperature converter.